test/requirements.txt
pykwalify
PyYAML
future
docopt
django
django-form-utils
docker-py
coveralls